requirements.txt
python-dotenv
langchain_ollama
langchain_text_splitters
langchain_chroma
pypdf
tqdm